# search-manage/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/freelancers/getDataWithVid/{vid}": {
            "get": {
                "description": "Retrieve freelancer data with a vector id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project/Freelancer"
                ],
                "summary": "Retrieve freelancer data with a vector id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vector Id",
                        "name": "vid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved freelancer data",
                        "schema": {
                            "$ref": "#/definitions/main.FreelancerList"
                        }
                    },
                    "400": {
                        "description": "Project ID missing in the path or is not a proper integer",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/freelancers/{id}": {
            "get": {
                "description": "Retrieve freelancer data with project id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project/Freelancer"
                ],
                "summary": "Retrieve freelancer data with project id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Freelancer Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved project data",
                        "schema": {
                            "$ref": "#/definitions/main.FreelancerList"
                        }
                    },
                    "400": {
                        "description": "Freelancer ID missing in the path or is not a proper integer",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/freelancers/{id}/find/projects": {
            "get": {
                "description": "Find freelancers clost to a given project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Find projects",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Freelancer Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found projects close to the input freelancer",
                        "schema": {
                            "$ref": "#/definitions/main.SearchProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Freelancer ID missing in the path",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Freelancer": {
            "type": "object",
            "properties": {
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "skills": {
                    "type": "string"
                }
            }
        },
        "main.FreelancerList": {
            "type": "object",
            "properties": {
                "freelancers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Freelancer"
                    }
                }
            }
        },
...
        "main.SearchProjectResponse": {
            "type": "object",
            "properties": {
                "D": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "numNeighbors": {
                    "type": "integer"
                },
                "numQueryVectors": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "main.VsDefaultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}